+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
  - The "align-" prefix was removed from `TextHorizontalAlignment` and `TextVerticalAlignment`. Either
    change `align-left` to `left` or qualify with `TextHorizontalAlignment.left`.
  - `img!"..."` is replaced by `@image-url("...")`
+ - `$children` is replaced by `@children`
 
 ### Added
  - title property to the Window element